refactor: split `SettingsSource` into `SettingsSource` and `EditableSettingsSource`
